biblio-zotero.el: add support for querying Zotero translation server
The package is modeled after biblio-doi.el. The main user-facing function is
`biblio-zotero-insert-bibtex' (aliased to `zotero-insert'), which inserts the
BibTeX entry returned by the Zotero server for the provided URL.

`biblio-zotero-forward-json' and `biblio-zotero-forward-bibtex' forward JSON or
BibTeX entry string to provided functions. `biblio-zotero-call-with-json'
and `biblio-zotero-call-with-bibtex' execute provided function in a buffer with
JSON/BibTeX entry.

Asynchronous communication with the Zotero server is done with `url-retrieve'
instead of `url-queue-retrieve' because we need to pass custom http headers.